crm114-20071201-BlameLittleStevie_VT-Ger-1515
* Made initial provisions to merge the crm tools with crm itself: added the
  statements cssmerge / cssdiff / cssinfo / cssanalyze / cssbackup / cssrestore / csscreate.

  These statements should suffice to cover the functionality found in the
  cssutil / osbf_util / cssmerge / cssdiff tools.

+++ src/testvectok.c
@@ -46,7 +46,7 @@ long user_trace = 0;
 
 long internal_trace = 0;
 
-long debug_countdown = 0;
+int debug_countdown = 0;
 
 long cmdline_break = 0;
